20200531_requirements.txt
astroid==2.4.1
get==2019.4.13
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
post==2019.4.13
public==2019.4.13
pylint==2.5.2
query-string==2019.4.13
request==2019.4.13
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Unidecode==1.1.1
wrapt==1.12.1
